(PC-13983)[ADAGE] refactor: move OfferType into types
