+++ adage-front/src/app/components/OffersInstantSearch/OffersSearch/OffersSearch.tsx
@@ -8,7 +8,8 @@ import { LEGACY_INITIAL_FACET_FILTERS, INITIAL_QUERY } from 'app/constants'
 import { FacetFiltersContext, AlgoliaQueryContext } from 'app/providers'
 import { FiltersContext } from 'app/providers/FiltersContextProvider'
 import { Filters } from 'app/types'
-import { Role, VenueFilterType } from 'utils/types'
+import { VenueFilterType } from 'app/types/offers'
+import { Role } from 'utils/types'
 
 import { populateFacetFilters } from '../utils'
